+++ integer/CHANGELOG.md
@@ -3,7 +3,7 @@
 ## 0.3.1 (WIP)
 
 - Add `count_ones()` and `count_zeros()` for `UBig`
-- Add `cubic()` for `UBig`
+- Add `cubic()` for `UBig` and `IBig`
 
 ## 0.3.0