+++ rational/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Impl `Sum` and `Product` traits for `RBig` and `Relaxed`.
 - Impl `dashu_base::Inverse` trait for `RBig` and `Relaxed`.
+- Add `cubic()` for `RBig` and `Relaxed`.
 
 ## 0.3.0 (Initial release)